refactor: SecurePrefsStore (Keystore AES/GCM) en lugar de EncryptedSharedPreferences y migración a Compose/OkHttp 5
- Nuevo SecurePrefsStore con clave AES-256/GCM en Android Keystore y migración transparente
- Reemplaza EncryptedSharedPreferences en NotesViewModel, BackupViewModel y SyncWorker
- Migra APIs deprecadas de Compose (Clipboard, PrimaryTabRow, ExposedDropdownMenuAnchorType)
- OkHttp 5: response.body no nullable
- Actualiza AGENTS.md (versiones, editor por bloques, IA llama.cpp)

+++ app/src/main/java/com/example/ui/BlockEditor.kt
@@ -129,7 +129,7 @@ fun BlockEditor(
                                         if (!longPressed) {
                                             val now = System.nanoTime()
                                             val elapsed = now - startTime
-                                            if (elapsed >= viewConfiguration.longPressTimeoutMillis.toLong() * 1_000_000) {
+                                            if (elapsed >= viewConfiguration.longPressTimeoutMillis * 1_000_000) {
                                                 longPressed = true
                                                 change.consume()
                                                 lastPosition = change.position
